For Clickpads, the finger must be pressing the touchpad for physical clicks.
This avoids clicks occuring when the clickpad is not being touched by a finger.
This happens for example with the Sony Vaio Pro, which has a very flexible chassis.

When sitting down with the laptop on your belly, and typing, this will otherwise trigger a clickpad physical click by
pressing the touchpad from under the chassis of the laptop. Also known as "belly-clicking".